feat: add empty pages(game, lobby, result) refactor: rename MainPage dir to Pages
