+++ src/component/Pages/GamePage.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const GamePage = (): JSX.Element => {
+  return (
+    <main>
+      <div>Game Page</div>
+    </main>
+  );
+};
+
+export default GamePage;
