+++ src/component/Pages/GameResultPage.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const GameResultPage = (): JSX.Element => {
+  return (
+    <main>
+      <div>Game result</div>
+    </main>
+  );
+};
+
+export default GameResultPage;
