+++ src/component/Pages/LobbyPage.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const LobbyPage = (): JSX.Element => {
+  return (
+    <main>
+      <div>Lobby Page</div>
+    </main>
+  );
+};
+
+export default LobbyPage;
